requirements.txt

streamlit>=1.27.0
pymupdf>=1.23.6
pdfplumber>=0.10.2
pytesseract>=0.3.10
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
nest-asyncio>=1.5.7
langchain-text-splitters>=0.0.1
langchain-community>=0.0.8
langgraph>=0.0.14
unstructured>=0.10.30
pdfminer.six==20221105
pypdf>=3.17.0
PyStemmer>=2.2.0
chromadb>=0.4.18
llama-index>=0.9.11
llama-index-embeddings-openai>=0.1.5
llama-index-vector-stores-chroma>=0.1.4
llama-index-retrievers-bm25>=0.1.1
llama-index-postprocessor-flag-embedding-reranker>=0.1.1
camelot-py[cv]>=0.11.0